RCF Expanded
The final version of the Recursive Categorical Framework is now published. This does not negate the original, but rather takes the same theorem's and reveals every one of the 16 new. This contains all operators, proofs, and expanded code snippets, alongside a few more frameworks. There is a json explanation of each section that allows for easy navigation.